Change max audio channels setting to a string
Turns out the legacy DB stored this as an ID to a value on the frontend
(e.g. 3 => "2.1"). We're just going to store the raw string of the
actual audio channel value.

I haven't hooked up a framework for new DB migrations, so this will
require a manual migration of the settings.json file for the new value.